// src/app/statistics/models/Spec.ts
export enum Spec {
  Restoration = 'Restoration',
  Holy = 'Holy',
  Discipline = 'Discipline',
  Mistweaver = 'Mistweaver',

  Protection = 'Protection',
  Guardian = 'Guardian',
  Blood = 'Blood',
  Vengeance = 'Vengeance',
  Brewmaster = 'Brewmaster',

  Elemental = 'Elemental',
  Balance = 'Balance',

  Enhancement = 'Enhancement',
  Havoc = 'Havoc',
  Feral = 'Feral',
  Arms = 'Arms',
  Fury = 'Fury',
  Frost = 'Frost',
  Unholy = 'Unholy',
  Affliction = 'Affliction',
  Demonology = 'Demonology',
  Destruction = 'Destruction',
  Subtlty = 'Subtlty',
  Assassination = 'Assassination',
  Outlaw = 'Outlaw',
  Retribution = 'Retribution',
  Windwalker = 'Windwalker',
  Fire = 'Fire',
  Arcane = 'Arcane',
  BeastMastery = 'Beast Mastery',
  Marksmanship = 'Marksmanship',
  Survival = 'Survival',
  Shadow = 'Shadow',
}